refactor(bridge-scene): move shop deep links out of the wearables domain
The marketplace lookup is neither a wearables nor an emotes concern, and
importing it from emotes made emotes depend on wearables. Give it its own
domain module; both import from there. No behaviour change.
